lecture-10-logging/logging_django/requirements.txt
django~=4.2.10
djangorestframework~=3.14.0